Proper handling of roles in the UI and many performance improvements.

+++ coderadar-core/src/main/java/io/reflectoring/coderadar/analyzer/domain/MetricValue.java
@@ -1,5 +1,6 @@
 package io.reflectoring.coderadar.analyzer.domain;
 
+import io.reflectoring.coderadar.plugin.api.Finding;
 import java.util.List;
 import lombok.AllArgsConstructor;
 import lombok.Data;
@@ -9,7 +10,7 @@ import lombok.Data;
 @AllArgsConstructor
 public class MetricValue {
   private final String name;
-  private final long value;
+  private final int value;
   private final long commitId;
   private final long fileId;
